src/services/triton_service/requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.10.0; python_version >= '3.9'
certifi==2025.8.3; python_version >= '3.7'
charset-normalizer==3.4.3; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
fastapi==0.116.1; python_version >= '3.8'
filelock==3.19.1; python_version >= '3.9'
fsspec==2025.9.0; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
huggingface-hub==0.34.4; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
iniconfig==2.1.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
mpmath==1.3.0
networkx==3.5; python_version >= '3.11'
numpy==1.23.5; python_version >= '3.11'
packaging==25.0; python_version >= '3.8'
pluggy==1.6.0; python_version >= '3.9'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-settings==2.10.1; python_version >= '3.9'
pygments==2.19.2; python_version >= '3.8'
pytest==8.4.1; python_version >= '3.9'
pytest-asyncio==1.1.0; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
regex==2025.9.1; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
safetensors==0.6.2; python_version >= '3.9'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.47.3; python_version >= '3.9'
sympy==1.14.0; python_version >= '3.9'
tokenizers==0.22.0; python_version >= '3.9'
torch==2.7.1; python_full_version >= '3.9.0'
tqdm==4.67.1; python_version >= '3.7'
transformers==4.56.0; python_full_version >= '3.9.0'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
uvicorn==0.35.0; python_version >= '3.9'
nvidia-pytriton==0.7.0